Add retries while sending commands to the background context
Apparently, sometimes the browser-action context can be started
before the background is ready to receive messages, the simple
retry mechanism introduced solves this problem.

A way to reproduce this is to leave a view opened in a tab, restart
the browser so that this view is the first displayed tab, when
communicating to the background, it will fail because the background
is not ready.